requirements.txt
langchain>=0.2.0
langchain-google-genai>=2.0.0
langchain-core>=0.2.0
langsmith>=0.1.0
python-dotenv>=1.0.0
